% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{control_plan}
\alias{control_plan}
\title{PARAMETRA control_plan data}
\format{
A data frame.
}
\usage{
control_plan
}
\description{
Information about disease control programs and plans implemented in different countries
}
\details{
Columns:
\describe{\item{country}{Country or region where data was collected}}
\describe{\item{country_disease_status}{Disease status classification}}
\describe{\item{filled_by}{Data entry person identifier}}
\describe{\item{host}{Host species}}
\describe{\item{host_age}{Age or age group of host}}
\describe{\item{host_breed}{Specific breed of host}}
\describe{\item{host_other}{Additional host characteristics}}
\describe{\item{host_type}{Type of host}}
\describe{\item{id}{Parameter identifier}}
\describe{\item{last_updated}{Date of last control plan update}}
\describe{\item{pathogen}{Name of the infectious agent}}
\describe{\item{plan_coverage}{Geographical scope of the plan}}
\describe{\item{plan_implemented}{Whether a control plan is implemented}}
\describe{\item{plan_type}{Type of control plan}}
\describe{\item{plan_voluntary_compulsary}{Voluntary or compulsory plan}}
\describe{\item{ref}{Reference: source DOI or URL}}
\describe{\item{ref_last_access}{Date reference was automatically checked for the last time}}
\describe{\item{ref_short}{Short reference: source author and year}}
\describe{\item{ref_status}{Reference status in the last aoutomatic check}}
\describe{\item{remarks}{Additional notes or comments}}
\describe{\item{year}{Year of publication}}
}
\keyword{datasets}
