% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ControlPlan.R
\docType{data}
\name{ControlPlan}
\alias{ControlPlan}
\title{Control Plan Parameters Table}
\format{
A data frame with 19 variables:
\describe{
\item{Country}{Country where the control plan is implemented}
\item{Pathogen}{Name of the infectious agent}
\item{Host}{Host species}
\item{Host type}{Type of host}
\item{Host breed}{Specific breed of host}
\item{Host Age (group)}{Age or age group of hosts}
\item{Host other characteristics}{Additional host characteristics}
\item{Control programme in place}{Whether a control programme exists}
\item{Type of programme}{Voluntary or compulsary programme}
\item{Coverage (regional/national)}{Geographical scope of the program}
\item{Country status for the disease}{Disease status classification}
\item{Control plan type}{Type of control measures implemented}
\item{Last updated}{Date of last program update}
\item{Remarks}{Additional notes or comments}
\item{Reference}{Citation for the source information}
\item{Year}{Year of implementation or publication}
\item{Link}{URL to program documentation}
\item{Filled in by:}{Data entry person identifier}
\item{ParameterType}{Category of parameter}
}
}
\source{
Compiled from various official sources and publications (see "Reference" column)
}
\usage{
ControlPlan
}
\description{
A dataset containing information about disease control programs and plans
implemented in different countries for various animal pathogens.
}
\keyword{datasets}
