% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cost_biosecurity}
\alias{cost_biosecurity}
\title{PARAMETRA cost_biosecurity data}
\format{
A data frame.
}
\usage{
cost_biosecurity
}
\description{
Information about prices of products needed to implement biosecurity measures
}
\details{
Columns:
\describe{\item{biosecurity_measure}{Category of measure}}
\describe{\item{country}{Country or region where data was collected}}
\describe{\item{filled_by}{Data entry person identifier}}
\describe{\item{filled_date}{Data entry date}}
\describe{\item{id}{Parameter identifier}}
\describe{\item{last_access}{URL last access date}}
\describe{\item{lifespan}{Expected lifespan of the product if applicable and based on typical usage}}
\describe{\item{parameter}{Parameter measured}}
\describe{\item{price}{Price of the product available online without VAT}}
\describe{\item{price_by_event}{Total cost per event to implement the biosecurity measure without VAT}}
\describe{\item{product_name}{Commercial name of the product}}
\describe{\item{product_type}{Specific type of product needed to implement the biosecurity measure}}
\describe{\item{qty_by_event}{Quantity of the product used per event to implement the biosecurity measure}}
\describe{\item{ref}{Reference: source DOI or URL}}
\describe{\item{ref_last_access}{Date reference was automatically checked for the last time}}
\describe{\item{ref_short}{Short reference: source author and year}}
\describe{\item{ref_status}{Reference status in the last aoutomatic check}}
\describe{\item{remarks}{Additional notes or comments}}
\describe{\item{transmission_pathway}{Transmission pathway targeted by this biosecurity measure}}
\describe{\item{unit_needed}{Unit of the product needed per event to implement the biosecurity measure (e.g. gram, liter, unit)}}
\describe{\item{year_cost}{Year the cost was included}}
}
\keyword{datasets}
