% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{diagnostic_test}
\alias{diagnostic_test}
\title{PARAMETRA diagnostic_test data}
\format{
A data frame.
}
\usage{
diagnostic_test
}
\description{
Diagnostic test parameters such as test specificity and sensitivity
}
\details{
Columns:
\describe{\item{filled_by}{Data entry person identifier}}
\describe{\item{filled_date}{Data entry date}}
\describe{\item{host}{Host species}}
\describe{\item{host_age}{Age or age group of host}}
\describe{\item{host_breed}{Specific breed of host}}
\describe{\item{host_other}{Additional host characteristics}}
\describe{\item{host_type}{Type of host}}
\describe{\item{id}{Parameter identifier}}
\describe{\item{lower_cl}{Lower bound of confidence/credible interval}}
\describe{\item{parameter}{Parameter measured}}
\describe{\item{pathogen}{Name of the infectious agent}}
\describe{\item{ref}{Reference: source DOI or URL}}
\describe{\item{ref_last_access}{Date reference was automatically checked for the last time}}
\describe{\item{ref_short}{Short reference: source author and year}}
\describe{\item{ref_status}{Reference status in the last aoutomatic check}}
\describe{\item{remarks}{Additional notes or comments}}
\describe{\item{sample_type}{Type of biological sample used}}
\describe{\item{study_type}{Study type}}
\describe{\item{test_type}{Type of diagnostic test used}}
\describe{\item{upper_cl}{Upper bound of confidence/credible interval}}
\describe{\item{value}{Numeric value of the parameter}}
\describe{\item{year}{Year of publication}}
}
\keyword{datasets}
