% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{infectious_latent_incuba_period}
\alias{infectious_latent_incuba_period}
\title{PARAMETRA infectious_latent_incuba_period data}
\format{
A data frame.
}
\usage{
infectious_latent_incuba_period
}
\description{
Infectious period, latent period, incubation period, shape and other disease period parameters
}
\details{
Columns:
\describe{\item{estimation_method}{Method used for estimation}}
\describe{\item{estimation_type}{Statistical approach used (Frequentist/Bayesian)}}
\describe{\item{filled_by}{Data entry person identifier}}
\describe{\item{filled_date}{Data entry date}}
\describe{\item{host}{Host species}}
\describe{\item{host_age}{Age or age group of host}}
\describe{\item{host_breed}{Specific breed of host}}
\describe{\item{host_other}{Additional host characteristics}}
\describe{\item{host_type}{Type of host}}
\describe{\item{id}{Parameter identifier}}
\describe{\item{if_other}{Additional parameter specifications}}
\describe{\item{inoculated_contact}{Inoculation of contact status}}
\describe{\item{lower_cl}{Lower bound of confidence/credible interval}}
\describe{\item{parameter}{Parameter measured}}
\describe{\item{pathogen}{Name of the infectious agent}}
\describe{\item{ref}{Reference: source DOI or URL}}
\describe{\item{ref_last_access}{Date reference was automatically checked for the last time}}
\describe{\item{ref_short}{Short reference: source author and year}}
\describe{\item{ref_status}{Reference status in the last aoutomatic check}}
\describe{\item{remarks}{Additional notes or comments}}
\describe{\item{scale}{Scale of the study}}
\describe{\item{study_type}{Study type}}
\describe{\item{unit}{Units of measurement}}
\describe{\item{upper_cl}{Upper bound of confidence/credible interval}}
\describe{\item{vaccine}{Type of vaccine used if applicable}}
\describe{\item{value}{Numeric value of the parameter}}
\describe{\item{variant_strain}{Specific variant or strain}}
\describe{\item{year}{Year of publication}}
}
\keyword{datasets}
