% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{other}
\alias{other}
\title{PARAMETRA other data}
\format{
A data frame.
}
\source{
Compiled from various scientific publications; see the "ref" column.
}
\usage{
other
}
\description{
Additional parameters and information related to various infectious diseases in animal populations that do not fit into other specific categories
}
\details{
Columns:
\describe{
\item{estimation_type}{Statistical approach used (Frequentist/Bayesian)}
\item{filled_by}{Data entry person identifier}
\item{filled_date}{Data entry date}
\item{host}{Host species}
\item{host_age}{Age or age group of host}
\item{host_breed}{Specific breed of host}
\item{host_other}{Additional host characteristics}
\item{host_type}{Type of host}
\item{id}{Parameter identifier}
\item{if_other}{Additional parameter specifications}
\item{lower_cl}{Lower bound of confidence/credible interval}
\item{model_type}{Model classification (e.g., SI, SIR)}
\item{parameter}{Parameter measured}
\item{pathogen}{Name of the infectious agent}
\item{ref}{Reference: source DOI or URL}
\item{ref_last_access}{Date reference was automatically checked for the last time}
\item{ref_short}{Short reference: source author and year}
\item{ref_status}{Reference status in the last aoutomatic check}
\item{remarks}{Additional notes or comments}
\item{study_type}{Study type}
\item{unit}{Units of measurement}
\item{upper_cl}{Upper bound of confidence/credible interval}
\item{vaccine}{Type of vaccine used if applicable}
\item{value}{Numeric value of the parameter}
\item{variant_strain}{Specific variant or strain}
\item{year}{Year of publication}
}
}
\keyword{datasets}
