% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pathogen_survival}
\alias{pathogen_survival}
\title{PARAMETRA pathogen_survival data}
\format{
A data frame.
}
\source{
Compiled from various scientific publications; see the "ref" column.
}
\usage{
pathogen_survival
}
\description{
Survival parameters in different materials and environmental conditions
}
\details{
Columns:
\describe{
\item{filled_by}{Data entry person identifier}
\item{filled_date}{Data entry date}
\item{id}{Parameter identifier}
\item{if_other}{Additional parameter specifications}
\item{material}{Type of material or medium where survival was measured}
\item{parameter}{Parameter measured}
\item{pathogen}{Name of the infectious agent}
\item{ref}{Reference: source DOI or URL}
\item{ref_last_access}{Date reference was automatically checked for the last time}
\item{ref_short}{Short reference: source author and year}
\item{ref_status}{Reference status in the last aoutomatic check}
\item{remarks}{Additional notes or comments}
\item{study_type}{Study type}
\item{variant_strain}{Specific variant or strain}
\item{year}{Year of publication}
}
}
\keyword{datasets}
