% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{regional_prevalence}
\alias{regional_prevalence}
\title{PARAMETRA regional_prevalence data}
\format{
A data frame.
}
\source{
Compiled from various scientific publications; see the "ref" column.
}
\usage{
regional_prevalence
}
\description{
Regional pathogen prevalence parameters at country or sub-division level
}
\details{
Columns:
\describe{
\item{country}{Country or region where data was collected}
\item{filled_by}{Data entry person identifier}
\item{host}{Host species}
\item{host_age}{Age or age group of host}
\item{host_breed}{Specific breed of host}
\item{host_other}{Additional host characteristics}
\item{host_type}{Type of host}
\item{id}{Parameter identifier}
\item{info_level}{Geographical level of information detail}
\item{parameter}{Parameter measured}
\item{pathogen}{Name of the infectious agent}
\item{ref}{Reference: source DOI or URL}
\item{ref_last_access}{Date reference was automatically checked for the last time}
\item{ref_short}{Short reference: source author and year}
\item{ref_status}{Reference status in the last aoutomatic check}
\item{remarks}{Additional notes or comments}
\item{sampling_purpose}{Purpose of the sampling conducted}
\item{study_type}{Study type}
\item{time_period}{Time period of the study}
\item{year}{Year of publication}
}
}
\keyword{datasets}
