% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RegionalPrevalence.R
\docType{data}
\name{RegionalPrevalence}
\alias{RegionalPrevalence}
\title{Regional Prevalence Parameters Table}
\format{
A data frame with 18 variables:
\describe{
\item{country}{Country or region where data was collected}
\item{parameter}{Parameter measured}
\item{sampling_purpose}{Purpose of the sampling conducted}
\item{info_level}{Geographical level of information detail}
\item{pathogen}{Name of the infectious agent}
\item{host}{Host species}
\item{host_type}{Type of host}
\item{host_breed}{Specific breed of host}
\item{host_age}{Age or age group of hosts}
\item{host_other}{Additional host characteristics}
\item{time_period}{Time period of the study}
\item{study_type}{Study type}
\item{remarks}{Additional notes or comments}
\item{ref_short}{Short reference: source author and year}
\item{year}{Year of publication}
\item{ref}{Reference: source doi or url}
\item{filled_by}{Data entry person identifier}
\item{parameter_type}{Category of parameter}
}
}
\source{
Compiled from various scientific publications (see "ref" column)
}
\usage{
RegionalPrevalence
}
\description{
A dataset containing regional prevalence parameters for various pathogens
in animal populations at country or regional level.
}
\keyword{datasets}
