% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{control_plan}
\alias{control_plan}
\title{PARAMETRA control_plan data}
\format{
A data frame.
}
\source{
Compiled from various scientific publications; see the "ref" column.
}
\usage{
control_plan
}
\description{
Information about disease control programs and plans implemented in different countries
}
\details{
Columns:
\describe{
\item{country}{Country or region where data was collected}
\item{country_disease_status}{Disease status classification}
\item{filled_by}{Data entry person identifier}
\item{host}{Host species}
\item{host_age}{Age or age group of host}
\item{host_breed}{Specific breed of host}
\item{host_other}{Additional host characteristics}
\item{host_type}{Type of host}
\item{id}{Parameter identifier}
\item{last_updated}{Date of last control plan update}
\item{pathogen}{Name of the infectious agent}
\item{plan_coverage}{Geographical scope of the plan}
\item{plan_implemented}{Whether a control plan is implemented}
\item{plan_type}{Type of control plan}
\item{plan_voluntary_compulsary}{Voluntary or compulsory plan}
\item{ref}{Reference: source DOI or URL}
\item{ref_last_access}{Date reference was automatically checked for the last time}
\item{ref_short}{Short reference: source author and year}
\item{ref_status}{Reference status in the last aoutomatic check}
\item{remarks}{Additional notes or comments}
\item{year}{Year of publication}
}
}
\keyword{datasets}
