% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cost_biosecurity}
\alias{cost_biosecurity}
\title{PARAMETRA cost_biosecurity data}
\format{
A data frame.
}
\source{
Compiled from various scientific publications; see the "ref" column.
}
\usage{
cost_biosecurity
}
\description{
Information about prices of products needed to implement biosecurity measures
}
\details{
Columns:
\describe{
\item{biosecurity_measure}{Category of measure}
\item{country}{Country or region where data was collected}
\item{filled_by}{Data entry person identifier}
\item{filled_date}{Data entry date}
\item{id}{Parameter identifier}
\item{last_access}{URL last access date}
\item{lifespan}{Expected lifespan of the product if applicable and based on typical usage}
\item{parameter}{Parameter measured}
\item{price}{Price of the product available online without VAT}
\item{price_by_event}{Total cost per event to implement the biosecurity measure without VAT}
\item{product_name}{Commercial name of the product}
\item{product_type}{Specific type of product needed to implement the biosecurity measure}
\item{qty_by_event}{Quantity of the product used per event to implement the biosecurity measure}
\item{ref}{Reference: source DOI or URL}
\item{ref_last_access}{Date reference was automatically checked for the last time}
\item{ref_short}{Short reference: source author and year}
\item{ref_status}{Reference status in the last aoutomatic check}
\item{remarks}{Additional notes or comments}
\item{transmission_pathway}{Transmission pathway targeted by this biosecurity measure}
\item{unit_needed}{Unit of the product needed per event to implement the biosecurity measure (e.g. gram, liter, unit)}
\item{year_cost}{Year the cost was included}
}
}
\keyword{datasets}
