% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{crossref}
\alias{crossref}
\title{PARAMETRA Crossref data}
\format{
A data frame.
}
\source{
Compiled from various scientific publications; see the "ref" column.
}
\usage{
crossref
}
\description{
Metadata extracted from CrossRef using the rcrossref package for scientific publications referenced in the parametra_long dataset
}
\details{
Columns:
\describe{
\item{id}{Parameter identifier}
\item{ref}{Reference: source DOI or URL}
\item{ref_last_access}{Date reference was automatically checked for the last time}
\item{ref_status}{Reference status in the last aoutomatic check}
}
}
\keyword{datasets}
