% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{diagnostic_test}
\alias{diagnostic_test}
\title{PARAMETRA diagnostic_test data}
\format{
A data frame.
}
\source{
Compiled from various scientific publications; see the "ref" column.
}
\usage{
diagnostic_test
}
\description{
Diagnostic test parameters such as test specificity and sensitivity
}
\details{
Columns:
\describe{
\item{filled_by}{Data entry person identifier}
\item{filled_date}{Data entry date}
\item{host}{Host species}
\item{host_age}{Age or age group of host}
\item{host_breed}{Specific breed of host}
\item{host_other}{Additional host characteristics}
\item{host_type}{Type of host}
\item{id}{Parameter identifier}
\item{lower_cl}{Lower bound of confidence/credible interval}
\item{parameter}{Parameter measured}
\item{pathogen}{Name of the infectious agent}
\item{ref}{Reference: source DOI or URL}
\item{ref_last_access}{Date reference was automatically checked for the last time}
\item{ref_short}{Short reference: source author and year}
\item{ref_status}{Reference status in the last aoutomatic check}
\item{remarks}{Additional notes or comments}
\item{sample_type}{Type of biological sample used}
\item{study_type}{Study type}
\item{test_type}{Type of diagnostic test used}
\item{upper_cl}{Upper bound of confidence/credible interval}
\item{value}{Numeric value of the parameter}
\item{year}{Year of publication}
}
}
\keyword{datasets}
