% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{parametra_crossref}
\alias{parametra_crossref}
\title{PARAMETRA Crossref metadata}
\format{
A data frame.
}
\source{
Metadata retrieved from Crossref for DOI references used in PARAMETRA.
}
\usage{
parametra_crossref
}
\description{
Crossref metadata for DOI references used in PARAMETRA.
}
\keyword{datasets}
