% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pathogen_survival}
\alias{pathogen_survival}
\title{PARAMETRA pathogen_survival data}
\format{
A data frame.
}
\usage{
pathogen_survival
}
\description{
Survival parameters in different materials and environmental conditions
}
\details{
Columns:
\describe{\item{filled_by}{Data entry person identifier}}
\describe{\item{filled_date}{Data entry date}}
\describe{\item{id}{Parameter identifier}}
\describe{\item{if_other}{Additional parameter specifications}}
\describe{\item{material}{Type of material or medium where survival was measured}}
\describe{\item{parameter}{Parameter measured}}
\describe{\item{pathogen}{Name of the infectious agent}}
\describe{\item{ref}{Reference: source DOI or URL}}
\describe{\item{ref_last_access}{Date reference was automatically checked for the last time}}
\describe{\item{ref_short}{Short reference: source author and year}}
\describe{\item{ref_status}{Reference status in the last aoutomatic check}}
\describe{\item{remarks}{Additional notes or comments}}
\describe{\item{study_type}{Study type}}
\describe{\item{variant_strain}{Specific variant or strain}}
\describe{\item{year}{Year of publication}}
}
\keyword{datasets}
