% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{regional_prevalence}
\alias{regional_prevalence}
\title{PARAMETRA regional_prevalence data}
\format{
A data frame.
}
\usage{
regional_prevalence
}
\description{
Regional pathogen prevalence parameters at country or sub-division level
}
\details{
Columns:
\describe{\item{country}{Country or region where data was collected}}
\describe{\item{filled_by}{Data entry person identifier}}
\describe{\item{host}{Host species}}
\describe{\item{host_age}{Age or age group of host}}
\describe{\item{host_breed}{Specific breed of host}}
\describe{\item{host_other}{Additional host characteristics}}
\describe{\item{host_type}{Type of host}}
\describe{\item{id}{Parameter identifier}}
\describe{\item{info_level}{Geographical level of information detail}}
\describe{\item{parameter}{Parameter measured}}
\describe{\item{pathogen}{Name of the infectious agent}}
\describe{\item{ref}{Reference: source DOI or URL}}
\describe{\item{ref_last_access}{Date reference was automatically checked for the last time}}
\describe{\item{ref_short}{Short reference: source author and year}}
\describe{\item{ref_status}{Reference status in the last aoutomatic check}}
\describe{\item{remarks}{Additional notes or comments}}
\describe{\item{sampling_purpose}{Purpose of the sampling conducted}}
\describe{\item{study_type}{Study type}}
\describe{\item{time_period}{Time period of the study}}
\describe{\item{year}{Year of publication}}
}
\keyword{datasets}
